updates for rexray

+++ scripts/tb.sh
@@ -52,10 +52,13 @@ echo SECONDMDMIP    = "${SECONDMDMIP}"
 echo CLUSTERINSTALL = "${CLUSTERINSTALL}"
 #echo "Number files in SEARCH PATH with EXTENSION:" $(ls -1 "${SEARCHPATH}"/*."${EXTENSION}" | wc -l)
 truncate -s 100GB ${DEVICE}
-yum install numactl libaio wget -y
+yum install numactl libaio wget unzip -y
 # install docker experimental
 wget -nv https://get.docker.com/rpm/1.7.0/centos-7/RPMS/x86_64/docker-engine-1.7.0-1.el7.centos.x86_64.rpm -O /tmp/docker.rpm
 yum install /tmp/docker.rpm -y
+systemctl stop docker
+rm -Rf /var/lib/docker
+sed -i -e "s/^OPTIONS=/#OPTIONS=/g" /etc/sysconfig/docker
 wget -nv https://experimental.docker.com/builds/Linux/x86_64/docker-latest -O /bin/docker
 systemctl restart docker
 # install rexray
@@ -69,7 +72,8 @@ EnvironmentFile=/etc/environment
 ExecStart=/bin/rexray --daemon
 ExecReload=/bin/kill -HUP $MAINPID
 KillMode=process
-Restart=on-failure
+Restart=always
+RestartSec=1
 [Install]
 WantedBy=docker.service' >> /usr/lib/systemd/system/rexray.service
 echo 'GOSCALEIO_ENDPOINT=https://192.168.50.12/api' >> /etc/environment
@@ -82,8 +86,12 @@ echo 'GOSCALEIO_STORAGEPOOL=pool1' >> /etc/environment
 systemctl daemon-reload
 systemctl start rexray.service
 cd /vagrant
-wget -nv ftp://ftp.emc.com/Downloads/ScaleIO/ScaleIO_RHEL6_Download.zip -O ScaleIO_RHEL6_Download.zip
-unzip -o ScaleIO_RHEL6_Download.zip -d /vagrant/scaleio/
+
+if [ ! -e "ScaleIO_RHEL6_Download.zip" ]; then
+  wget -nv ftp://ftp.emc.com/Downloads/ScaleIO/ScaleIO_RHEL6_Download.zip -O ScaleIO_RHEL6_Download.zip
+  unzip -o ScaleIO_RHEL6_Download.zip -d /vagrant/scaleio/
+fi
+
 cd /vagrant/scaleio/ScaleIO_1.32_RHEL7_Download
 
 if [ "${CLUSTERINSTALL}" == "True" ]; then
